[herd,asl] Fix unsound XOR optimization
For XOR, the optimization when the two values are equal (or compare
returns 0) is to return V.zero. However, the actual result might be
different for different scalar types. Thus V.zero should not be used
in all the cases.

An example is the ASLConstant module, when XOR-ing two bitvectors, the
result should be a zeroed bitvector of the same length, rather than the
S_Int Z.zero.

Removing the optimization fixes the failed test.